+++ x/vm/keeper/keeper.go
@@ -103,7 +103,6 @@ func NewKeeper(
 	fmk types.FeeMarketKeeper,
 	consensusKeeper types.ConsensusParamsKeeper,
 	erc20Keeper types.Erc20Keeper,
-	evmChainID uint64,
 	tracer string,
 ) *Keeper {
 	// ensure evm module account is set
@@ -121,12 +120,6 @@ func NewKeeper(
 	storeKeys := make(map[string]storetypes.StoreKey)
 	for _, k := range keys {
 		storeKeys[k.Name()] = k
-	}
-
-	// set global chain config
-	ethCfg := types.DefaultChainConfig(evmChainID)
-	if err := types.SetChainConfig(ethCfg); err != nil {
-		panic(err)
 	}
 
 	// NOTE: we pass in the parameter space to the CommitStateDB in order to use custom denominations for the EVM operations